refactor: extract workspace duplication outside CreateWorkspacePage file
